+++ src/typeparser/variables.ts
@@ -1,5 +1,4 @@
 import * as vscode from "vscode";
-
 import { nameType } from "./types/typeClasses";
 
 export class VariableParser {
@@ -7,21 +6,21 @@ export class VariableParser {
     if (text == undefined) return [];
     const output: nameType[] = [];
     const allLines = text.split("\n");
-    const variableRegex =
-      /((^(\s+)?declare\s+(metadata\s+|netread\s+|netwrite\s+|persistent\s+){0,1}(.*?)\s)|(#Const\s))\w+/gm;
+    const regex =
+      /(\s+)?declare\s+(metadata\s+|netread\s+|netwrite\s+|persistent\s+){0,1}(.*?)\s+(\w+)\s*(for\s+\w+){0,1}\s*(=|;)/g;
 
     for (const line in allLines) {
-      const lineText = allLines[line];
-      const regex = lineText.match(variableRegex);
-      if (regex == null) continue;
-      const match = regex[0];
-      const result = this.parseVariable(match);
-      if (result) {
+      let m;
+      while ((m = regex.exec(allLines[line])) !== null) {
+        // This is necessary to avoid infinite loops with zero-width matches
+        if (m.index === regex.lastIndex) {
+          regex.lastIndex++;
+        }
+
         const i = Number.parseInt(line);
-        const start = allLines[i].indexOf(result.name, 9 + result.type.length);
-        const range = new vscode.Range(i, start, i, start + result.name.length);
-        result.range = range;
-        output.push(result);
+        const start = allLines[i].indexOf(m[4], 9 + m[3].length);
+        const range = new vscode.Range(i, start, i, start + m[4].length);
+        output.push({ name: m[4], type: m[3], range: range });
       }
     }
 
@@ -59,32 +58,6 @@ export class VariableParser {
     return output;
   }
 
-  parseVariable(variableString: string): nameType | undefined {
-    const mArray = variableString.replace(/^\s+/, "").split(/\s/);
-
-    const name = mArray.pop() ?? "";
-    const type = mArray.pop() ?? "";
-
-    const storageKeywords = [
-      "declare",
-      "#Const",
-      "persistent",
-      "metadata",
-      "netwrite",
-      "netread",
-    ];
-
-    if (storageKeywords.includes(type) == false) {
-      // if type matches [], it must be array
-      if (type.match(/\[(.*)?\]/g)) {
-        // var types = type.match(/\w+(?!\[\])?/g);
-        // arrayTypes[variablesNames.indexOf(name)] = types;
-      }
-      return { name: name, type: type };
-    }
-    return;
-  }
-
   parseForeach(text: string): nameType[] {
     if (text == undefined) return [];
     const output: nameType[] = [];
@@ -95,15 +68,21 @@ export class VariableParser {
       foreachRegex.lastIndex = -1;
       const mArray = foreachRegex.exec(match);
       if (Array.isArray(mArray) && mArray.length == 4) {
+        const keyName = mArray[1];
         const name = mArray[2];
         const type = mArray[3];
         const offset = text.indexOf(match.trim());
         const lines = text.slice(0, offset).split("\n");
         const allLines = text.split("\n");
         const i = lines.length - 1;
+        
+        if (keyName) {
+          const start2 = allLines[i].indexOf(keyName);
+          const range2 = new vscode.Range(i, start2, i, start2 + keyName.length);
+          output.push({ name: keyName, type: "Integer", range: range2 });
+        }
         const start = allLines[i].indexOf(name);
         const range = new vscode.Range(i, start, i, start + name.length);
-
         output.push({ name: name, type: type, range: range });
       }
     }